+++ NEWS.md
@@ -1,7 +1,32 @@
+# plm 2.2-4
+
+* ptransform (internal function): check balancedness before pseries index is
+  removed (fixes some spurious bug, e.g. when package tibble is used).
+* exported/registered again in NAMESPACE after export/registration lost in plm 2.0-0:
+  fixef.pggls, Math.pseries, Ops.pseries, Complex.pseries and deprecated
+  methods/function formula.dynformula, print.dynformula, pvcovHC.
+* Ops.pseries: use of is.vector() was too strict, now uses is.atomic() with
+  taking care for additional data types.
+* pwaldtest:
+   * non-exported function wald() now exported as method pwaldtest.pgmm.
+   * for all plm models use approach via
+     crossprod(solve(vcov(x), coefs), coefs)), not (tss-ssr)/(ssr/df2)
+     anymore.
+   * method for pvcm models now allows for pvcm's "within" specification,
+     returning a data.frame with test results for each regression.
+* pcdtest.pseries: NA values in input variable are now removed before any
+  further processing. A warning is issued in case of NA removal.
+* mtest, sargan, pwaldtest, piest, aneweytest: added for each a string for
+  alternative hypothesis.
+* Dependencies: removed package 'clubSandwich' from 'Suggests' as it was
+  removed from CRAN (archived).
+
 # plm 2.2-3
-* IGNORE_RDIFF_BEGIN/END added on tests and man pages
+
+* IGNORE_RDIFF_BEGIN/END added on tests and man pages.
 
 # plm 2.2-1
+
 * purtest:
    * tests now support unbalanced panel data, where applicable.
    * gained argument 'ips.stat' to select statistic for IPS test,
@@ -19,28 +44,30 @@
 * index.pindex: fixed bug when individual index variable is called "group".
 
 ## Minor items:
-
    * print.fixef: respects / forwards arguments supplied to generic print method.
    * DESCRIPTION/Dependencies: package 'urca' added to "Suggests".
    * Grunfeld data doc: URL for overview of various Grunfeld data sets updated to
      https://eeecon.uibk.ac.at/~zeileis/grunfeld/. 
 
 # plm 2.2-0
-* methods for plm.list where not exported, now exported.
+
+* methods for plm.list were not exported, now exported.
 * lagt is changed so that it can deal with time factors which
   cannot be coerced to numeric (ex "1950-54", "1955-59", ...).
 * cortab was not exported, now exported.
-* pvcm failed for random effect models when there are some NA coefs for
+* pvcm failed for random effect models when there are some NA coefficients for
   some individual level OLS regressions, fixed.
 
 # plm 2.1-0
+
 * problems with vignettes fixed (full text was in italics).
-* in test_Estimators, L256, tolerance lowered to 1E-04.
+* in test file 'test_Estimators.R', L256, tolerance lowered to 1E-04.
 
 # plm 2.0-2
+
 * vcovXX.pcce functions exported again (export was lost in plm 2.0-0).
-* summary.pcce gained argument 'vcov', summary.pcce object carries robust vcov in
-  element 'rvcov'.
+* summary.pcce gained argument 'vcov', summary.pcce object carries robust vcov
+  in element 'rvcov'.
 * vignettes switched from bookdown::html_document2 to html_vignette.
 
 # plm 2.0-1